backend/requirements.txt
# Core web framework
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Google Cloud client libraries
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0

# Async support
aiohttp==3.8.5
asyncio==3.4.3

# Utilities
python-dotenv==1.0.0
pydantic==2.3.0
structlog==23.1.0

# Development dependencies
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.1.0
mypy==1.5.1 